chore: migrate deps to unconfirmedlabs org and refactor events
- Rename author/license holder LLC -> Inc. across all modules
- Switch BPS dependency interest_bps -> bps (unconfirmedlabs/bps)
- Point partyos/share deps at unconfirmedlabs org
- Republish: update package IDs and upgrade cap in Published.toml/Move.lock
- Normalize split_value to u16; simplify WalrusData event fields to blob_id
- Clean up event schemas (rename *CreditRoleEvent -> *CreditRoleAssignedEvent,
  drop redundant count/duration fields, add published_by)
- Refactor composition/recording/release/deal/track accordingly

Pre-existing working-tree changes, committed as the base for metadata slimming.

+++ tests/release_tests.move
@@ -39,7 +39,7 @@ fun test_track(ctx: &mut TxContext): track::Track {
 }
 
 /// Helper to create a disc with n tracks that share splits evenly.
-fun test_disc_with_n_tracks(n: u64, split_bps: u64, ctx: &mut TxContext): disc::Disc {
+fun test_disc_with_n_tracks(n: u64, split_bps: u16, ctx: &mut TxContext): disc::Disc {
     let mut tracks = vector[];
     n.do!(|_| {
         tracks.push_back(track::new_for_testing<CompositionShare, RecordingShare, V>(
